Added the Ach. Schema columns with the courses bin count
